1. Изменил сумму займа с Integer на double.

+++ entity/src/main/java/ru/devufa/debt/entity/Dept.java
@@ -26,7 +26,7 @@ public class Dept extends EntityWithId{
         Сумма займа
      */
     @Column(nullable = false)
-    private Integer count;
+    private double count;
     /*
         Валюта
      */
@@ -66,11 +66,11 @@ public class Dept extends EntityWithId{
         this.borrower = borrower;
     }
 
-    public Integer getCount() {
+    public double getCount() {
         return count;
     }
 
-    public void setCount(Integer count) {
+    public void setCount(double count) {
         this.count = count;
     }
